Cleaned up classes.py
set_var handles both assignment and reassignment, so I've removed def_var. Making room for when I went to add user defined types, I've simplified the code for checking them and added conversion. Conversion added primarily for the special case of a float needing to be converted to an int.