Fixed boolean support
Wanted to handle booleans separately in the interpreter to make sure that if someone returned a string value 'true'/'false' it would not be correct for a function that should return an actual boolean literal.

Meanwhile, I created the boolean literal class that I should have made a while ago and added parser support for returning instances of it.