Fixed variable definition and assignment
There were two issues.
1. Variables could not be properly defined without being initalized with values, which was my intention.
2. Setting variables would not check if they already existed in the current scope before type checking, therefore completely ignoring types during reassignment.

I resolved these by adding a different kind of assignment to the parser that provided None as its expression, updated the interpreter for that, and added the variable definition method back to the environment class.

The set_var method is bulky and needs rework, will work on it soon. Send over a pull request if you think you can help.