Scope fixes
I removed the newlines skip in the parser and that messed some stuff up so I added it back haha

Anyways, let's get into this. Recursive stuff is handled way better now, and scope actually works. I also added numbers to the environments to see how deeply nested the scopes are in testing.